refactor(webview): rename module inline.rs -> mount.rs
