+++ src/webview.rs
@@ -1,14 +1,14 @@
 //! In-process webview feature: CEF render wiring and the window.ozma Tier 1
 //! back-channel (render), OSC mount/unmount of inline webviews (osc), and
-//! inline webviews rendered into the terminal text flow (inline). Aggregated
+//! inline webviews rendered into the terminal text flow (mount). Aggregated
 //! behind OzmuxWebviewPlugin.
 
-pub(crate) mod inline;
+pub(crate) mod mount;
 pub(crate) mod osc;
 pub(crate) mod render;
 
 use bevy::prelude::*;
-use inline::WebviewPlugin;
+use mount::WebviewPlugin;
 use osc::OscPlugin;
 use render::RenderPlugin;
